+++ lib/active_encryption.rb
@@ -2,7 +2,7 @@
 
 require 'active_encryption/version'
 require 'active_encryption/configuration'
-require 'active_encryption/encryptable'
+require 'active_encryption/adapters'
 
 # The top-level ActiveEncryption module isolates the gem from the host
 # application.