+++ lib/active_encryption/adapters.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require 'active_encryption/encryptable'
+require 'active_encryption/adapters/active_record'