+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+- Add ActiveRecord adapter
 - Add ActiveEncryption::Encryptable concern to (en|de)crypt object attributes
 - Add ActiveEncryption::Encryptor to abstract ActiveSupport::MessageEncryptor
 - Add DSL to configure the gem